Refactor Chat page, moving methods and constants and methods to Chat.util file
